HW7 Part B: Advanced graphics done! :)
Forget to change the containerConnector back.

Added drag and drop. (drag will call selectPiece and drop will cal
selectPiece or selectSquare accordingly) Notice that only the current
turn player can drag the pieces that can perform a move, otherwise the
pieces can not be dragged.

Added animation and sound but it does not support drap and drop
obviously, so if perform a move by dnd, then no animation or sound will
be called.

app engine url:
http://1-7-0-dot-banqi-smg.appspot.com/