Game refactored! XD
1. The state is optimized, the cell (previous square) is represent as
C0-C31. So for C0: obj, obj can be a piece or an empty cell or null
(invisible). It saves a lot of effort later :)
2. All classes relevant to the new state are updated.
3. All tests are also updated and passed.
4. Delete some useless classes. (e.g. CapturePiece.java)
5. All images are updated and some new added. Thx to my girlfriend
joyce's effort :)
6. Add a new feature: as a piece is selected, all possible target cells
or pieces will be highlighted.
7. Fixed a performence problem in BanqiGraphics, the game will not be
slower and slower as before. (Previous I forgot to clear the board panel
on each update... so the original images are still hanging around
beneath which cause a major performance hit...)

Two bugs are known for now.
Animation bug: when the selected piece begins to move, it already
appears in the destination... I've identified the cause: right after the
animation is called, the view will updated the board first, then the
animation is performed and completed... Though I'm still thinking a
clean solution without messing the codes and making it more
complicated...

Dnd bug: if a piece is dragged to a illegal place and dropped, it will
disappear... Though after legally clicking other pieces, it will
appears...

Note this should be V2 and the appengine url is: